refactor(backend): complete 5-wave backend modularization
Restructure the backend from a monolithic routes/mod.rs (1395 lines) into a modular
provider architecture where all providers flow through the Provider trait uniformly.

Wave 1 - Provider trait enhancement + polymorphic dispatch:
- Add normalize_response_for_openai/anthropic to Provider trait
- Introduce ProviderMap (HashMap<ProviderId, Arc<dyn Provider>>)
- Replace match chains with map lookup via state.providers.get()
- Add as_any() for downcasting trait objects to concrete types

Wave 2 - Encapsulate provider-specific state:
- Move copilot_token_cache + device_pending into CopilotProvider
- Move qwen_device_pending into QwenProvider
- Remove 3 fields from AppState, access via downcast pattern

Wave 3 - Split routes/mod.rs into sub-modules:
- Extract routes/state.rs (AppState, SessionInfo, UserKiroCreds)
- Extract routes/pipeline.rs (shared pipeline helpers)
- Extract routes/openai.rs (chat_completions_handler)
- Extract routes/anthropic.rs (anthropic_messages_handler)
- Slim routes/mod.rs to router setup only

Wave 4 - Unify pipeline logic:
- Extract ProviderRouting, resolve_provider_routing()
- Extract build_kiro_credentials(), read_config()
- Both handlers use identical provider dispatch code path

Wave 5 - Complete KiroProvider:
- Fully implement KiroProvider with all 4 Provider trait methods
- Wire converters, streaming, token counting into provider
- Remove inlined Kiro pipeline from both handlers
- All providers (including Kiro) now flow through Provider trait

Adding a new provider now requires only: create provider file, add ProviderId
variant, register in build_provider_map(). No changes to routes or handlers.

All 779 tests pass. E2E port defaults updated to match docker config.
