+++ backend/src/web_ui/mod.rs
@@ -1,10 +1,10 @@
 pub mod api_keys;
 pub mod config_api;
 pub mod config_db;
-pub mod model_registry;
-pub mod model_registry_handlers;
 pub mod copilot_auth;
 pub mod google_auth;
+pub mod model_registry;
+pub mod model_registry_handlers;
 pub mod provider_oauth;
 pub mod provider_priority;
 pub mod qwen_auth;
@@ -100,7 +100,10 @@ pub fn web_ui_routes(state: AppState) -> Router {
         .merge(config_api::user_routes())
         .merge(crate::guardrails::api::guardrails_routes())
         .nest("/admin/mcp", crate::mcp::api::mcp_admin_routes())
-        .nest("/admin/models", model_registry_handlers::model_registry_routes())
+        .nest(
+            "/admin/models",
+            model_registry_handlers::model_registry_routes(),
+        )
         .layer(axum::middleware::from_fn(google_auth::admin_middleware))
         .layer(axum::middleware::from_fn(google_auth::csrf_middleware))
         .layer(axum::middleware::from_fn_with_state(
